Also separated FormatExtensionList out of Geometry (and made Attribute and Group subtypes), converting it from a struct to class with private members
Using debug new on Windows (in debug mode, of course)

Fixed leak in ShaderData
