+++ librtt/Core/Rtt_Allocator.h
@@ -95,7 +95,12 @@ void operator delete[]( void* p ) throw();
 
 	#ifdef __cplusplus
 
-	#define Rtt_NEW( pUnused, T )				new T
+	#if defined(Rtt_WIN_ENV) && defined(Rtt_DEBUG)
+		#define Rtt_NEW( pUnused, T )			new ( _CLIENT_BLOCK, __FILE__, __LINE__) T
+	#else
+		#define Rtt_NEW( pUnused, T )			new T
+	#endif
+
 	#define Rtt_DELETE( p )						delete (p)
 
 #ifdef Rtt_PLACEMENT_NEW
