+++ librtt/Display/Rtt_ShaderData.cpp
@@ -62,6 +62,11 @@ ShaderData::~ShaderData()
 		fProxy->DetachUserdata();
 	}
 
+	for ( int i = 0; i < kNumData; i++ )
+	{
+		Rtt_DELETE( fUniformData[i] );
+	}
+
     Rtt_FREE( fExtraSpace );
 }
 
@@ -81,6 +86,8 @@ ShaderData::QueueRelease( DisplayObject *observer )
 	for ( int i = 0; i < kNumData; i++ )
 	{
 		observer->QueueRelease( fUniformData[i] );
+
+		fUniformData[i] = NULL;
 	}
 	observer->QueueRelease( fProxy );
 }
