+++ librtt/Renderer/Rtt_GLGeometry.cpp
@@ -12,6 +12,7 @@
 
 #include "Renderer/Rtt_GLGeometry.h"
 
+#include "Renderer/Rtt_FormatExtensionList.h"
 #include "Renderer/Rtt_Geometry_Renderer.h"
 #include "Renderer/Rtt_GL.h"
 
@@ -476,7 +477,7 @@ GLGeometry::Update( CPUResource* resource )
     
     if (gainedExtension)
     {
-        Rtt_ASSERT( extensionList->groupCount > 0 );
+        Rtt_ASSERT( extensionList->GetGroupCount() > 0 );
         
         fInstancesAllocated = 0;
     }
@@ -746,7 +747,7 @@ GLGeometry::Bind()
 }
 
 static void
-BindExtensionAttribute( const Geometry::ExtensionAttribute& attribute, GLuint attributeIndex, size_t size, GLbyte* start, U32 offsetExtra )
+BindExtensionAttribute( const FormatExtensionList::Attribute& attribute, GLuint attributeIndex, size_t size, GLbyte* start, U32 offsetExtra )
 {
     GLenum type = GL_FLOAT;
     
@@ -781,8 +782,8 @@ GLGeometry::ResolveVertexFormat( const FormatExtensionList * list, U32 vertexSiz
     
     for ( auto iter = FormatExtensionList::AllGroups( list ); !iter.IsDone(); iter.Advance() )
     {
-        const Geometry::ExtensionGroup* group = iter.GetGroup();
-        const Geometry::ExtensionAttribute* first = iter.GetAttribute();
+        const FormatExtensionList::Group* group = iter.GetGroup();
+        const FormatExtensionList::Attribute* first = iter.GetAttribute();
         GLbyte* start = NULL;
         U32 offsetExtra = 0;
         size_t stride;
