+++ platform/windows/Corona.Rtt.Library/Corona.Rtt.Library.Win32.vcxproj
@@ -161,6 +161,7 @@
     <ClCompile Include="..\..\..\librtt\Input\Rtt_ReadOnlyInputDeviceCollection.cpp" />
     <ClCompile Include="..\..\..\librtt\Renderer\Rtt_CommandBuffer.cpp" />
     <ClCompile Include="..\..\..\librtt\Renderer\Rtt_CPUResource.cpp" />
+    <ClCompile Include="..\..\..\librtt\Renderer\Rtt_FormatExtensionList.cpp" />
     <ClCompile Include="..\..\..\librtt\Renderer\Rtt_FrameBufferObject.cpp" />
     <ClCompile Include="..\..\..\librtt\Renderer\Rtt_Geometry_Renderer.cpp">
       <ObjectFileName Condition="'$(Configuration)|$(Platform)'=='Release.App|Win32'">$(IntDir)Renderer\</ObjectFileName>
@@ -467,6 +468,7 @@
     <ClInclude Include="..\..\..\librtt\Input\Rtt_ReadOnlyInputDeviceCollection.h" />
     <ClInclude Include="..\..\..\librtt\Renderer\Rtt_CommandBuffer.h" />
     <ClInclude Include="..\..\..\librtt\Renderer\Rtt_CPUResource.h" />
+    <ClInclude Include="..\..\..\librtt\Renderer\Rtt_FormatExtensionList.h" />
     <ClInclude Include="..\..\..\librtt\Renderer\Rtt_FrameBufferObject.h" />
     <ClInclude Include="..\..\..\librtt\Renderer\Rtt_Geometry_Renderer.h" />
     <ClInclude Include="..\..\..\librtt\Renderer\Rtt_GeometryPool.h" />
