+++ platform/windows/Corona.Rtt.Library/Corona.Rtt.Library.Win32.vcxproj.filters
@@ -942,6 +942,9 @@
     <ClCompile Include="..\..\..\librtt\Display\Rtt_TextureResourceCaptureAdapter.cpp">
       <Filter>librtt\Display</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\..\librtt\Renderer\Rtt_FormatExtensionList.cpp">
+      <Filter>librtt\Renderer</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\external\ALmixer\luaal.h">
@@ -1894,6 +1897,9 @@
     </ClInclude>
     <ClInclude Include="..\..\..\librtt\Display\Rtt_TextureResourceCaptureAdapter.h">
       <Filter>librtt\Display</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\librtt\Renderer\Rtt_FormatExtensionList.h">
+      <Filter>librtt\Renderer</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
